refactor: replace BytesUtils with ByteUtils across the codebase
