+++ clients/src/test/java/org/apache/kafka/common/utils/BytesTest.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.common.utils;
 
-import org.apache.kafka.common.utils.internals.BytesUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 
 import org.junit.jupiter.api.Test;
 
@@ -36,14 +36,14 @@ public class BytesTest {
     public void testIncrement() {
         byte[] input = new byte[]{(byte) 0xAB, (byte) 0xCD, (byte) 0xFF};
         byte[] expected = new byte[]{(byte) 0xAB, (byte) 0xCE, (byte) 0x00};
-        Bytes output = BytesUtils.increment(Bytes.wrap(input));
+        Bytes output = ByteUtils.increment(Bytes.wrap(input));
         assertArrayEquals(output.get(), expected);
     }
 
     @Test
     public void testIncrementUpperBoundary() {
         byte[] input = new byte[]{(byte) 0xFF, (byte) 0xFF, (byte) 0xFF};
-        assertThrows(IndexOutOfBoundsException.class, () -> BytesUtils.increment(Bytes.wrap(input)));
+        assertThrows(IndexOutOfBoundsException.class, () -> ByteUtils.increment(Bytes.wrap(input)));
     }
 
     @Test
@@ -66,7 +66,7 @@ public class BytesTest {
         map.put(key5, val);
 
         Bytes prefix = key1;
-        Bytes prefixEnd = BytesUtils.increment(prefix);
+        Bytes prefixEnd = ByteUtils.increment(prefix);
 
         Comparator<? super Bytes> comparator = map.comparator();
         final int result = comparator == null ? prefix.compareTo(prefixEnd) : comparator.compare(prefix, prefixEnd);
@@ -114,7 +114,7 @@ public class BytesTest {
     }
 
     private int cmp(String l, String r) {
-        return BytesUtils.BYTES_LEXICO_COMPARATOR.compare(
+        return ByteUtils.BYTES_LEXICO_COMPARATOR.compare(
             l.getBytes(StandardCharsets.UTF_8),
             r.getBytes(StandardCharsets.UTF_8));
     }
