+++ streams/src/main/java/org/apache/kafka/streams/state/internals/DualColumnFamilyAccessor.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.streams.state.internals;
 
 import org.apache.kafka.common.utils.Bytes;
-import org.apache.kafka.common.utils.internals.BytesUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.errors.ProcessorStateException;
 import org.apache.kafka.streams.state.internals.RocksDBStore.DBAccessor;
@@ -271,7 +271,7 @@ class DualColumnFamilyAccessor extends AbstractColumnFamilyAccessor {
         // RocksDB's JNI interface does not expose getters/setters that allow the
         // comparator to be pluggable, and the default is lexicographic, so it's
         // safe to just force lexicographic comparator here for now.
-        private final Comparator<byte[]> comparator = BytesUtils.BYTES_LEXICO_COMPARATOR;
+        private final Comparator<byte[]> comparator = ByteUtils.BYTES_LEXICO_COMPARATOR;
 
         private final String storeName;
         private final RocksIterator iterNewFormat;
@@ -401,7 +401,7 @@ class DualColumnFamilyAccessor extends AbstractColumnFamilyAccessor {
     // RocksDB's JNI interface does not expose getters/setters that allow the
     // comparator to be pluggable, and the default is lexicographic, so it's
     // safe to just force lexicographic comparator here for now.
-        private final Comparator<byte[]> comparator = BytesUtils.BYTES_LEXICO_COMPARATOR;
+        private final Comparator<byte[]> comparator = ByteUtils.BYTES_LEXICO_COMPARATOR;
         private final byte[] rawLastKey;
         private final boolean forward;
         private final boolean toInclusive;
