+++ streams/src/main/java/org/apache/kafka/streams/state/internals/RocksDBRangeIterator.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.streams.state.internals;
 
 import org.apache.kafka.common.utils.Bytes;
-import org.apache.kafka.common.utils.internals.BytesUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.streams.KeyValue;
 
 import org.rocksdb.RocksIterator;
@@ -28,7 +28,7 @@ class RocksDBRangeIterator extends RocksDbIterator {
     // RocksDB's JNI interface does not expose getters/setters that allow the
     // comparator to be pluggable, and the default is lexicographic, so it's
     // safe to just force lexicographic comparator here for now.
-    private final Comparator<byte[]> comparator = BytesUtils.BYTES_LEXICO_COMPARATOR;
+    private final Comparator<byte[]> comparator = ByteUtils.BYTES_LEXICO_COMPARATOR;
     private final byte[] rawLastKey;
     private final boolean forward;
     private final boolean toInclusive;
